spotify-playlist-generator/requirements.txt
python-dotenv==1.0.0
openai==0.27.8
ipython==8.14.0
Flask==2.3.3
virtualenv==20.24.3
tiktoken==0.4.0
spotipy==2.23.0